+++ src/debug_tool/debug_tool_host.h
@@ -44,7 +44,7 @@ struct AlgorithmResourceBinding {
 
 struct AlgorithmDescriptorValue {
   std::string descriptor_name;
-  float scalar_value{0.0f};
+  double scalar_value{0.0};
 };
 
 struct AlgorithmPipelineStageSubmission {
